+++ src/app/(admin)/admin/inventory/page.tsx
@@ -10,6 +10,8 @@ import { usePostUploadImg } from '@/service/shared/shared.query';
 import Loading from '@/components/shared/Loading';
 import Image from 'next/image';
 import Sidebar from '@/components/Admin/Sidebar';
+import InventoryManagementModal from '@/components/Admin/InventoryManagementModal';
+import AddInventoryModal from '@/components/Admin/AddInventoryModal';
 
 interface ProductInfo {
   id: number;
@@ -47,13 +49,6 @@ const InventoryManagement = () => {
     useState<InventoryData | null>(null);
   const [isInventoryModalOpen, setIsInventoryModalOpen] = useState(false);
   const [isAddInventoryModalOpen, setIsAddInventoryModalOpen] = useState(false);
-  const [newInventory, setNewInventory] = useState({
-    productId: 0,
-    imgURL: '',
-    expiration: '',
-  });
-  const [selectedFile, setSelectedFile] = useState<File | null>(null);
-  const [previewImage, setPreviewImage] = useState<string>('');
 
   const { data, isLoading, refetch } = useGetAdminInventory();
   const createInventoryMutation = useCreateAdminInventory();
@@ -80,36 +75,21 @@ const InventoryManagement = () => {
 
   const handleAddInventory = () => {
     if (!selectedInventoryData) return;
-    setNewInventory({
-      productId: selectedInventoryData.product.id,
-      imgURL: '',
-      expiration: '',
-    });
-    setSelectedFile(null);
-    setPreviewImage('');
     setIsAddInventoryModalOpen(true);
   };
 
-  const handleFileChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-    const file = e.target.files?.[0];
-    if (file) {
-      setSelectedFile(file);
-      // 미리보기 이미지 생성
-      const reader = new FileReader();
-      reader.onload = (e) => {
-        setPreviewImage(e.target?.result as string);
-      };
-      reader.readAsDataURL(file);
-    }
-  };
-
-  const handleSaveInventory = async () => {
+  const handleSaveInventory = async (data: {
+    productId: number;
+    imgURL: string;
+    expiration: string;
+    selectedFile: File | null;
+  }) => {
     try {
       let imageUrl = '/images/barcode.jpg';
 
-      if (selectedFile) {
+      if (data.selectedFile) {
         const formData = new FormData();
-        formData.append('image', selectedFile);
+        formData.append('image', data.selectedFile);
 
         const uploadResult = await uploadImgMutation.mutateAsync(formData);
         if (uploadResult?.data?.imgURL) {
@@ -118,9 +98,9 @@ const InventoryManagement = () => {
       }
 
       await createInventoryMutation.mutateAsync({
-        productId: newInventory.productId,
+        productId: data.productId,
         imgURL: imageUrl,
-        expiration: newInventory.expiration,
+        expiration: data.expiration,
       });
 
       setIsAddInventoryModalOpen(false);
@@ -156,9 +136,7 @@ const InventoryManagement = () => {
         productId,
         inventoryId,
       });
-      alert('재고가 성공적으로 삭제되었습니다.');
 
-      // 데이터를 새로고침하고 현재 선택된 상품의 업데이트된 정보를 모달에 반영
       const refreshedData = await refetch();
       if (selectedInventoryData && refreshedData.data?.data) {
         const updatedInventoryData = refreshedData.data.data.find(
@@ -326,224 +304,23 @@ const InventoryManagement = () => {
         </main>
       </div>
 
-      {/* 재고 관리 모달 */}
-      {isInventoryModalOpen && selectedInventoryData && (
-        <div className="fixed inset-0 z-50 flex items-center justify-center p-4 bg-black bg-opacity-50">
-          <div className="bg-white rounded-lg shadow-xl max-w-4xl w-full max-h-[90vh] overflow-y-auto">
-            <div className="p-6">
-              <div className="flex items-center justify-between mb-4">
-                <h2 className="text-xl font-bold">
-                  재고 관리 - {selectedInventoryData.product.title}
-                </h2>
-                <button
-                  onClick={() => setIsInventoryModalOpen(false)}
-                  className="text-gray-500 hover:text-gray-700"
-                >
-                  ✕
-                </button>
-              </div>
+      {/* 새로운 모달 컴포넌트들 */}
+      <InventoryManagementModal
+        isOpen={isInventoryModalOpen}
+        onClose={() => setIsInventoryModalOpen(false)}
+        selectedInventoryData={selectedInventoryData}
+        onAddInventory={handleAddInventory}
+        onDeleteInventory={handleDeleteInventory}
+        isDeleting={deleteInventoryMutation.isPending}
+      />
 
-              <div className="mb-4">
-                <button
-                  onClick={handleAddInventory}
-                  className="px-4 py-2 text-sm text-white bg-blue-600 rounded-md hover:bg-blue-700"
-                >
-                  + 재고 추가
-                </button>
-              </div>
-
-              <div className="overflow-x-auto">
-                <table className="min-w-full divide-y divide-gray-200">
-                  <thead className="bg-gray-50">
-                    <tr>
-                      <th className="px-6 py-3 text-xs font-medium tracking-wider text-left text-gray-500 uppercase">
-                        ID
-                      </th>
-                      <th className="px-6 py-3 text-xs font-medium tracking-wider text-left text-gray-500 uppercase">
-                        이미지
-                      </th>
-                      <th className="px-6 py-3 text-xs font-medium tracking-wider text-left text-gray-500 uppercase">
-                        판매 상태
-                      </th>
-                      <th className="px-6 py-3 text-xs font-medium tracking-wider text-left text-gray-500 uppercase">
-                        사용 상태
-                      </th>
-                      <th className="px-6 py-3 text-xs font-medium tracking-wider text-left text-gray-500 uppercase">
-                        만료일
-                      </th>
-                      <th className="px-6 py-3 text-xs font-medium tracking-wider text-left text-gray-500 uppercase">
-                        생성일
-                      </th>
-                      <th className="px-6 py-3 text-xs font-medium tracking-wider text-left text-gray-500 uppercase">
-                        액션
-                      </th>
-                    </tr>
-                  </thead>
-                  <tbody className="bg-white divide-y divide-gray-200">
-                    {selectedInventoryData.inventories.map(
-                      (inventory: InventoryItem) => (
-                        <tr key={inventory.id}>
-                          <td className="px-6 py-4 text-sm text-gray-900 whitespace-nowrap">
-                            {inventory.id}
-                          </td>
-                          <td className="px-6 py-4 whitespace-nowrap">
-                            <div className="relative w-10 h-10">
-                              <Image
-                                src={inventory.imgURL}
-                                alt="재고 이미지"
-                                fill
-                                className="object-cover rounded"
-                              />
-                            </div>
-                          </td>
-                          <td className="px-6 py-4 whitespace-nowrap">
-                            <span
-                              className={`inline-flex px-2 py-1 text-xs font-semibold rounded-full ${
-                                inventory.isSold
-                                  ? 'bg-red-100 text-red-800'
-                                  : 'bg-green-100 text-green-800'
-                              }`}
-                            >
-                              {inventory.isSold ? '판매완료' : '판매가능'}
-                            </span>
-                          </td>
-                          <td className="px-6 py-4 whitespace-nowrap">
-                            <span
-                              className={`inline-flex px-2 py-1 text-xs font-semibold rounded-full ${
-                                inventory.isUse
-                                  ? 'bg-gray-100 text-gray-800'
-                                  : 'bg-blue-100 text-blue-800'
-                              }`}
-                            >
-                              {inventory.isUse ? '사용완료' : '사용가능'}
-                            </span>
-                          </td>
-                          <td className="px-6 py-4 text-sm text-gray-900 whitespace-nowrap">
-                            {new Date(
-                              inventory.expiration
-                            ).toLocaleDateString()}
-                          </td>
-                          <td className="px-6 py-4 text-sm text-gray-900 whitespace-nowrap">
-                            {new Date(inventory.createdAt).toLocaleDateString()}
-                          </td>
-                          <td className="px-6 py-4 text-sm font-medium whitespace-nowrap">
-                            <button
-                              onClick={() => {
-                                handleDeleteInventory(
-                                  inventory.id,
-                                  selectedInventoryData?.product.id
-                                );
-                                console.log(selectedInventoryData);
-                              }}
-                              disabled={deleteInventoryMutation.isPending}
-                              className="text-red-600 hover:text-red-900 disabled:text-gray-400 disabled:cursor-not-allowed"
-                            >
-                              {deleteInventoryMutation.isPending
-                                ? '삭제 중...'
-                                : '삭제'}
-                            </button>
-                          </td>
-                        </tr>
-                      )
-                    )}
-                  </tbody>
-                </table>
-              </div>
-
-              {selectedInventoryData.inventories.length === 0 && (
-                <div className="py-8 text-center">
-                  <div className="text-gray-500">재고가 없습니다.</div>
-                </div>
-              )}
-            </div>
-          </div>
-        </div>
-      )}
-
-      {/* 재고 추가 모달 */}
-      {isAddInventoryModalOpen && (
-        <div className="fixed inset-0 z-50 flex items-center justify-center p-4 bg-black bg-opacity-50">
-          <div className="w-full max-w-md bg-white rounded-lg shadow-xl">
-            <div className="p-6">
-              <div className="flex items-center justify-between mb-4">
-                <h2 className="text-xl font-bold">재고 추가</h2>
-                <button
-                  onClick={() => setIsAddInventoryModalOpen(false)}
-                  className="text-gray-500 hover:text-gray-700"
-                >
-                  ✕
-                </button>
-              </div>
-
-              <div className="space-y-4">
-                <div>
-                  <label className="block mb-2 text-sm font-medium text-gray-700">
-                    이미지 업로드
-                  </label>
-                  <input
-                    type="file"
-                    accept="image/*"
-                    onChange={handleFileChange}
-                    className="w-full px-3 py-2 border border-gray-300 rounded-md focus:outline-none focus:ring-2 focus:ring-blue-500"
-                  />
-                  {previewImage && (
-                    <div className="mt-2">
-                      <div className="relative w-20 h-20">
-                        <Image
-                          src={previewImage}
-                          alt="미리보기"
-                          fill
-                          className="object-cover rounded"
-                        />
-                      </div>
-                    </div>
-                  )}
-                  <p className="mt-1 text-xs text-gray-500">
-                    이미지를 업로드하지 않으면 기본 바코드 이미지가 사용됩니다.
-                  </p>
-                </div>
-
-                <div>
-                  <label className="block mb-2 text-sm font-medium text-gray-700">
-                    만료일
-                  </label>
-                  <input
-                    type="date"
-                    value={newInventory.expiration}
-                    onChange={(e) =>
-                      setNewInventory({
-                        ...newInventory,
-                        expiration: e.target.value,
-                      })
-                    }
-                    className="w-full px-3 py-2 border border-gray-300 rounded-md focus:outline-none focus:ring-2 focus:ring-blue-500"
-                    required
-                  />
-                </div>
-
-                <div className="flex space-x-2">
-                  <button
-                    onClick={handleSaveInventory}
-                    disabled={
-                      !newInventory.expiration ||
-                      createInventoryMutation.isPending
-                    }
-                    className="flex-1 px-4 py-2 text-sm text-white bg-blue-600 rounded-md hover:bg-blue-700 disabled:bg-gray-400 disabled:cursor-not-allowed"
-                  >
-                    {createInventoryMutation.isPending ? '저장 중...' : '저장'}
-                  </button>
-                  <button
-                    onClick={() => setIsAddInventoryModalOpen(false)}
-                    className="flex-1 px-4 py-2 text-sm text-gray-700 bg-gray-200 rounded-md hover:bg-gray-300"
-                  >
-                    취소
-                  </button>
-                </div>
-              </div>
-            </div>
-          </div>
-        </div>
-      )}
+      <AddInventoryModal
+        isOpen={isAddInventoryModalOpen}
+        onClose={() => setIsAddInventoryModalOpen(false)}
+        onSave={handleSaveInventory}
+        productId={selectedInventoryData?.product.id || 0}
+        isSaving={createInventoryMutation.isPending}
+      />
     </div>
   );
 };
